add PersonalTab
